+++ src/atoms/index.js
@@ -1,6 +1,5 @@
 // @flow
 
-export { default as Button } from './Button';
-export { default as ButtonQR } from './ButtonQR';
+export { default as InvoiceQR } from './InvoiceQR';
 export { default as Small } from './Small';
 export { default as Text } from './Text';