+++ src/components/BadgerButton/index.js
@@ -1,5 +1,0 @@
-// @flow
-
-import BadgerButton from './BadgerButton';
-
-export default BadgerButton;